final_project/requirements.txt
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlalchemy==2.0.48
boto3>=1.28.0
dbt-postgres==1.10.0
